رفع اشکال

+++ ApiKarbord/Models/Public/Web_IDocHMini.cs
@@ -137,7 +137,7 @@
 
         public double? ArzRate { get; set; }
 
-        public long? AccSerialNumber { get; set; }
+        public int? AccSerialNumber { get; set; }
 
         public int? AccDocNo { get; set; }
